testkit: split helpers and persistence suite
