+++ packages/minitsis-testkit/src/index.ts
@@ -26,6 +26,8 @@ import {
   setBufferSize,
 } from '@minitsis/core';
 import type {Database} from 'minitsis-datastore';
+import {MapDB, wrapWithName, wrapWithNameAsync} from './helpers';
+import {addPersistenceTests} from './persistence-tests';
 
 export interface DatabaseHandle {
   db: Database;
@@ -36,57 +38,6 @@ export interface MinitsisTestAdapter {
   name: string;
   makeDatabase: () => Promise<DatabaseHandle>;
   makePersistentDatabase?: () => Promise<DatabaseHandle>;
-}
-
-// In-memory Database implementation used for most core semantics tests
-class MapDB implements Database {
-  private data: Map<string, Uint8Array> = new Map();
-
-  async set(key: string, value: Uint8Array): Promise<void> {
-    this.data.set(key, value);
-  }
-
-  async get(key: string): Promise<Uint8Array | null> {
-    return this.data.has(key) ? this.data.get(key)! : null;
-  }
-
-  async delete(key: string): Promise<void> {
-    this.data.delete(key);
-  }
-
-  async count(): Promise<number> {
-    return this.data.size;
-  }
-}
-
-function wrapWithName(
-  testFn: (testCase: TestCase) => void
-): (testCase: TestCase) => void {
-  const currentTestName = expect.getState().currentTestName;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  (testFn as any).testName = currentTestName;
-  return testFn;
-}
-
-function wrapWithNameAsync(
-  testFn: (testCase: TestCase) => Promise<void>
-): (testCase: TestCase) => Promise<void> {
-  const currentTestName = expect.getState().currentTestName;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  (testFn as any).testName = currentTestName;
-  return testFn;
-}
-
-async function usingDatabase(
-  factory: () => Promise<DatabaseHandle>,
-  fn: (db: Database) => Promise<void>
-): Promise<void> {
-  const handle = await factory();
-  try {
-    await fn(handle.db);
-  } finally {
-    await handle.cleanup?.();
-  }
 }
 
 export function runCommonMinitsisTests(adapter: MinitsisTestAdapter): void {
@@ -183,73 +134,7 @@ export function runCommonMinitsisTests(adapter: MinitsisTestAdapter): void {
       ).rejects.toThrow(Unsatisfiable);
     });
 
-    describe('persistence via adapter database', () => {
-      const countMarker = {value: 0};
-
-      const runWithAdapter = async (): Promise<{dbCount: number}> => {
-        await usingDatabase(adapter.makePersistentDatabase || adapter.makeDatabase, async db => {
-          const testFn = (testCase: TestCase) => {
-            countMarker.value += 1;
-            const choice = testCase.choice(BigInt(10000));
-            if (choice >= BigInt(8)) {
-              throw new Error('Choice is too high');
-            }
-          };
-
-          await expect(
-            runTest(100, 1234, db, false)(wrapWithName(testFn))
-          ).rejects.toThrow('Choice is too high');
-        });
-
-        const handle = await adapter.makePersistentDatabase?.();
-        if (!handle) {
-          return {dbCount: 0};
-        }
-        const count = await handle.db.count();
-        await handle.cleanup?.();
-        return {dbCount: count};
-      };
-
-      test('reuses stored failing example', async () => {
-        if (!adapter.makePersistentDatabase) {
-          test.skip('adapter does not support persistence');
-          return;
-        }
-
-        const first = await adapter.makePersistentDatabase();
-        try {
-          const testFn = (testCase: TestCase) => {
-            countMarker.value += 1;
-            const choice = testCase.choice(BigInt(10000));
-            if (choice >= BigInt(8)) {
-              throw new Error('Choice is too high');
-            }
-          };
-
-          await expect(
-            runTest(100, 1234, first.db, false)(wrapWithName(testFn))
-          ).rejects.toThrow('Choice is too high');
-
-          const initialCount = await first.db.count();
-          const prev = countMarker.value;
-
-          await expect(
-            runTest(100, 1234, first.db, false)(wrapWithName(testFn))
-          ).rejects.toThrow('Choice is too high');
-
-          const afterCount = await first.db.count();
-          expect(initialCount).toBeGreaterThanOrEqual(1);
-          expect(afterCount).toBe(initialCount);
-          expect(countMarker.value).toBe(prev + 2);
-        } finally {
-          await first.cleanup?.();
-        }
-      });
-
-      test('adapter database reachable', async () => {
-        await runWithAdapter();
-      });
-    });
+    addPersistenceTests(adapter);
 
     test('test_function_cache', async () => {
       const testFn = async (testCase: TestCase) => {
